backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
openai==1.57.4
pdfminer.six==20231228
python-docx==1.1.2
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
anyio==4.7.0
